Changed Matrix type element to double

+++ src/codes/ADT/Matrix.java
@@ -2,11 +2,11 @@ package codes.ADT;
 
 public class Matrix {
     // Konstruktor Matrix
-    public int[][] Mtrx;
+    public double[][] Mtrx;
     public int rows = 0;
     public int cols = 0;
     public Matrix(int rows , int cols) {
-        this.Mtrx = new int[rows][cols];
+        this.Mtrx = new double[rows][cols];
         this.rows = rows;
         this.cols = cols;
     }
